feat: add sports and change location from enum to string
While editing a migration file really defeats the point of migrations in the first place, it is important that the `add_google_oauth` migration occurs after the collections have been created. While the other obvious way around this would be to use custom prefixes to force this ordering, PocketBase does recommend using timestamps to work well with its migration tracking and so since we haven't had a release yet, this seems like the best alternative.